feat: Merge SuezData and SuezAsyncClient into SuezClient
