+++ src/pysuez/__init__.py
@@ -1,0 +1,13 @@
+from pysuez.models import (
+    AggregatedData,
+    AlertResult,
+    ContractResult,
+    DayDataResult,
+    InterventionResult,
+    LimestoneResult,
+    PriceResult,
+    QualityResult,
+)
+
+from pysuez.exception import PySuezConnexionError, PySuezDataError, PySuezError
+from pysuez.suez_client import SuezClient
