+++ src/pysuez/__main__.py
@@ -3,8 +3,7 @@ import asyncio
 import sys
 from datetime import datetime, timedelta
 
-from .async_client import SuezAsyncClient
-from .suez_data import SuezData
+from pysuez.suez_client import SuezClient
 
 
 async def main():
@@ -22,37 +21,36 @@ async def main():
 
     args = parser.parse_args()
 
-    async_client = SuezAsyncClient(args.username, args.password, args.counter_id)
-    if args.counter_id is None:
-        await async_client.counter_finder()
-    data = SuezData(async_client)
-
+    client = SuezClient(args.username, args.password, args.counter_id)
     try:
+        if args.counter_id is None:
+            await client.find_counter()
+
         if args.mode == "alerts":
             print("getting alerts")
-            alerts = await data.get_alerts()
+            alerts = await client.get_alerts()
             print("leak=", alerts.leak, ", consumption=", alerts.overconsumption)
         elif args.mode == "test":
-            print(await data.contract_data())
-            print(await data.get_alerts())
-            print(await data.get_price())
-            print(await data.get_interventions())
-            print(await data.get_quality())
-            print(await data.get_limestone())
-            print(await data.fetch_yesterday_data())
+            print(await client.contract_data())
+            print(await client.get_alerts())
+            print(await client.get_price())
+            print(await client.get_interventions())
+            print(await client.get_water_quality())
+            print(await client.get_limestone())
+            print(await client.fetch_yesterday_data())
             print(
-                await data.fetch_all_available(
+                await client.fetch_all_daily_data(
                     since=(datetime.now() - timedelta(weeks=4)).date()
                 )
             )
-            print(await data.fetch_all_deprecated_data())
+            print(await client.fetch_aggregated_data())
         else:
-            print(await data.fetch_all_deprecated_data())
+            print(await client.fetch_aggregated_data())
     except BaseException as exp:
         print(exp)
         return 1
     finally:
-        await async_client.close_session()
+        await client.close_session()
 
 
 if __name__ == "__main__":
