+++ src/pysuez/const.py
@@ -9,7 +9,7 @@ INFORMATION_ENDPOINT_INTERVENTION = _INFORMATION_ENDPOINT + "intervention/"
 INFORMATION_ENDPOINT_QUALITY = _INFORMATION_ENDPOINT + "quality/"
 INFORMATION_ENDPOINT_PRICE = _INFORMATION_ENDPOINT + "price/"
 INFORMATION_ENDPOINT_LIMESTONE = _INFORMATION_ENDPOINT + "limestone/"
-API_ENDPOINT_DAY_DATA = "/mon-compte-en-ligne/statJData/"
+API_ENDPOINT_DAILY_DATA = "/mon-compte-en-ligne/statJData/"
 API_CONSUMPTION_INDEX = "/public-api/contract/tile/consumption"
 API_ENDPOINT_MONTH_DATA = "/mon-compte-en-ligne/statMData/"
 API_HISTORY_CONSUMPTION = "/mon-compte-en-ligne/historique-de-consommation-tr"
